♻️ Refactor error system into modular architecture and remove activities
- Convert monolithic errors.rs into modular system:
  - errors/common.rs: AppError trait and ErrorSeverity enum
  - errors/pet.rs: Pet-specific error types with AppError implementation
  - errors/activity.rs: Example extension for activity errors
  - errors/mod.rs: Module orchestration and re-exports
- Create separate validation module:
  - validation/pet.rs: Pet validation functions
  - validation/mod.rs: Module orchestration
- Remove activity-related code:
  - Delete commands/activities.rs and database/activities.rs
  - Update module exports and command handlers
- Update imports to use new modular structure:
  - commands/pets.rs: Use separate validation module
  - lib.rs: Include validation module
- Maintain backward compatibility through re-exports
