--- src-tauri/src/database/activities.rs
@@ -1,1317 +1,0 @@
-use super::models::*;
-use anyhow::Result;
-use chrono::{DateTime, Utc};
-use sqlx::Row;
-
-impl super::PetDatabase {
-    /// Create a new activity
-    pub async fn create_activity(&self, activity_data: ActivityCreateRequest) -> Result<Activity> {
-        let now = Utc::now();
-        let activity_data_json = activity_data
-            .activity_data
-            .map(|data| serde_json::to_string(&data))
-            .transpose()?;
-
-        let result = sqlx::query(
-            r#"
-            INSERT INTO activities (pet_id, category, subcategory, title, description, activity_date, 
-                                  activity_data, cost, currency, location, mood_rating, created_at, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            "#
-        )
-        .bind(activity_data.pet_id)
-        .bind(activity_data.category.to_string())
-        .bind(&activity_data.subcategory)
-        .bind(&activity_data.title)
-        .bind(&activity_data.description)
-        .bind(activity_data.activity_date)
-        .bind(activity_data_json)
-        .bind(activity_data.cost)
-        .bind(&activity_data.currency)
-        .bind(&activity_data.location)
-        .bind(activity_data.mood_rating)
-        .bind(now)
-        .bind(now)
-        .execute(&self.pool)
-        .await?;
-
-        // Update FTS table
-        if let Some(activity_id) = Some(result.last_insert_rowid()) {
-            let _ = sqlx::query(
-                "INSERT INTO activities_fts(rowid, title, description, category, subcategory) 
-                 VALUES (?, ?, ?, ?, ?)",
-            )
-            .bind(activity_id)
-            .bind(&activity_data.title)
-            .bind(&activity_data.description)
-            .bind(activity_data.category.to_string())
-            .bind(&activity_data.subcategory)
-            .execute(&self.pool)
-            .await;
-        }
-
-        self.get_activity_by_id(result.last_insert_rowid()).await
-    }
-
-    /// Get activities with optional filtering
-    pub async fn get_activities(
-        &self,
-        pet_id: Option<i64>,
-        filters: Option<ActivityFilters>,
-        limit: Option<i64>,
-        offset: Option<i64>,
-    ) -> Result<ActivitySearchResult> {
-        let mut base_query = String::from(
-            "SELECT id, pet_id, category, subcategory, title, description, activity_date, 
-             activity_data, cost, currency, location, mood_rating, created_at, updated_at,
-             COUNT(*) OVER() as total_count FROM activities WHERE 1=1",
-        );
-        let mut bind_values = Vec::new();
-
-        // Pet filter
-        if let Some(pid) = pet_id {
-            base_query.push_str(" AND pet_id = ?");
-            bind_values.push(pid.to_string());
-        }
-
-        // Category filters
-        if let Some(filters) = &filters {
-            if let Some(categories) = &filters.categories {
-                let category_placeholders = categories
-                    .iter()
-                    .map(|_| "?")
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                base_query.push_str(&format!(" AND category IN ({category_placeholders})"));
-                for category in categories {
-                    bind_values.push(category.to_string());
-                }
-            }
-
-            // Date filters
-            if let Some(date_from) = filters.date_from {
-                base_query.push_str(" AND activity_date >= ?");
-                bind_values.push(date_from.to_rfc3339());
-            }
-            if let Some(date_to) = filters.date_to {
-                base_query.push_str(" AND activity_date <= ?");
-                bind_values.push(date_to.to_rfc3339());
-            }
-
-            // Cost filters
-            if let Some(min_cost) = filters.min_cost {
-                base_query.push_str(" AND cost >= ?");
-                bind_values.push(min_cost.to_string());
-            }
-            if let Some(max_cost) = filters.max_cost {
-                base_query.push_str(" AND cost <= ?");
-                bind_values.push(max_cost.to_string());
-            }
-        }
-
-        base_query.push_str(" ORDER BY activity_date DESC");
-
-        // Add pagination
-        if let Some(limit) = limit {
-            base_query.push_str(" LIMIT ?");
-            bind_values.push(limit.to_string());
-
-            if let Some(offset) = offset {
-                base_query.push_str(" OFFSET ?");
-                bind_values.push(offset.to_string());
-            }
-        }
-
-        // Build the final query with proper pagination
-        let mut final_query = String::from(
-            "SELECT a.id, a.pet_id, a.category, a.subcategory, a.title, a.description,
-                   a.activity_date, a.activity_data, a.cost, a.currency, a.location,
-                   a.mood_rating, a.created_at, a.updated_at,
-                   COUNT(*) OVER() as total_count
-            FROM activities a
-            WHERE 1=1",
-        );
-
-        // Add pet filter if provided
-        if let Some(pid) = pet_id {
-            final_query.push_str(" AND a.pet_id = ");
-            final_query.push_str(&pid.to_string());
-        }
-
-        final_query.push_str(" ORDER BY a.activity_date DESC");
-
-        // Add pagination
-        let limit_val = limit.unwrap_or(50);
-        let offset_val = offset.unwrap_or(0);
-        final_query.push_str(" LIMIT ");
-        final_query.push_str(&limit_val.to_string());
-        final_query.push_str(" OFFSET ");
-        final_query.push_str(&offset_val.to_string());
-
-        let rows = sqlx::query(&final_query).fetch_all(&self.pool).await?;
-
-        let mut activities = Vec::new();
-        let mut total_count = 0i64;
-
-        for row in rows {
-            if total_count == 0 {
-                total_count = row.try_get("total_count").unwrap_or(0);
-            }
-            activities.push(self.row_to_activity(&row).await?);
-        }
-
-        let has_more = if let Some(offset) = offset {
-            (offset + activities.len() as i64) < total_count
-        } else {
-            false
-        };
-
-        Ok(ActivitySearchResult {
-            activities,
-            total_count,
-            has_more,
-        })
-    }
-
-    /// Get activity by ID
-    pub async fn get_activity_by_id(&self, id: i64) -> Result<Activity> {
-        let row = sqlx::query("SELECT * FROM activities WHERE id = ?")
-            .bind(id)
-            .fetch_one(&self.pool)
-            .await?;
-
-        self.row_to_activity(&row).await
-    }
-
-    /// Update an activity
-    pub async fn update_activity(
-        &self,
-        id: i64,
-        activity_data: ActivityUpdateRequest,
-    ) -> Result<Activity> {
-        // Build dynamic update query
-        let mut updates = Vec::new();
-
-        // Collect updates and their values
-        let mut bind_values = Vec::new();
-
-        if let Some(title) = &activity_data.title {
-            updates.push("title = ?");
-            bind_values.push(title.clone());
-        }
-        if let Some(subcategory) = &activity_data.subcategory {
-            updates.push("subcategory = ?");
-            bind_values.push(subcategory.clone());
-        }
-        if let Some(description) = &activity_data.description {
-            updates.push("description = ?");
-            bind_values.push(description.clone());
-        }
-
-        // Use a simpler approach - build a single query with all updates
-        if !updates.is_empty() {
-            let sql = format!(
-                "UPDATE activities SET {}, updated_at = CURRENT_TIMESTAMP WHERE id = ?",
-                updates.join(", ")
-            );
-
-            // For simplicity, let's handle each field individually
-            sqlx::query(&sql)
-                .bind(activity_data.title.clone().unwrap_or_default())
-                .bind(activity_data.subcategory.clone().unwrap_or_default())
-                .bind(activity_data.description.clone().unwrap_or_default())
-                .bind(id)
-                .execute(&self.pool)
-                .await?;
-        }
-
-        self.get_activity_by_id(id).await
-    }
-
-    /// Delete an activity
-    pub async fn delete_activity(&self, id: i64) -> Result<()> {
-        sqlx::query("DELETE FROM activities WHERE id = ?")
-            .bind(id)
-            .execute(&self.pool)
-            .await?;
-        Ok(())
-    }
-
-    /// Search activities using basic text search (fallback when FTS is not available)
-    pub async fn search_activities(
-        &self,
-        search_query: &str,
-        pet_id: Option<i64>,
-        limit: Option<i64>,
-        offset: Option<i64>,
-    ) -> Result<ActivitySearchResult> {
-        // Use basic LIKE search as fallback when FTS is not available
-        let mut sql = String::from(
-            "SELECT a.id, a.pet_id, a.category, a.subcategory, a.title, a.description,
-                   a.activity_date, a.activity_data, a.cost, a.currency, a.location,
-                   a.mood_rating, a.created_at, a.updated_at,
-                   COUNT(*) OVER() as total_count
-            FROM activities a
-            WHERE (a.title LIKE ? OR a.description LIKE ? OR a.subcategory LIKE ?)",
-        );
-
-        if let Some(pet_id_val) = pet_id {
-            sql.push_str(" AND a.pet_id = ");
-            sql.push_str(&pet_id_val.to_string());
-        }
-
-        sql.push_str(" ORDER BY a.activity_date DESC LIMIT ? OFFSET ?");
-
-        let search_pattern = format!("%{search_query}%");
-        let rows = sqlx::query(&sql)
-            .bind(&search_pattern)
-            .bind(&search_pattern)
-            .bind(&search_pattern)
-            .bind(limit.unwrap_or(50))
-            .bind(offset.unwrap_or(0))
-            .fetch_all(&self.pool)
-            .await?;
-
-        let mut activities = Vec::new();
-        let mut total_count = 0i64;
-
-        for row in rows {
-            if total_count == 0 {
-                total_count = row.try_get("total_count").unwrap_or(0);
-            }
-            activities.push(self.row_to_activity(&row).await?);
-        }
-
-        let has_more = if let Some(offset) = offset {
-            (offset + activities.len() as i64) < total_count
-        } else {
-            false
-        };
-
-        Ok(ActivitySearchResult {
-            activities,
-            total_count,
-            has_more,
-        })
-    }
-
-    /// Create an activity attachment
-    pub async fn create_activity_attachment(
-        &self,
-        activity_id: i64,
-        file_path: String,
-        file_type: ActivityAttachmentType,
-        file_size: Option<i64>,
-        thumbnail_path: Option<String>,
-        metadata: Option<serde_json::Value>,
-    ) -> Result<ActivityAttachment> {
-        let now = Utc::now();
-        let metadata_json = metadata
-            .map(|data| serde_json::to_string(&data))
-            .transpose()?;
-
-        let result = sqlx::query(
-            r#"
-            INSERT INTO activity_attachments (activity_id, file_path, file_type, file_size, thumbnail_path, metadata, created_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?)
-            "#
-        )
-        .bind(activity_id)
-        .bind(&file_path)
-        .bind(file_type.to_string())
-        .bind(file_size)
-        .bind(&thumbnail_path)
-        .bind(metadata_json)
-        .bind(now)
-        .execute(&self.pool)
-        .await?;
-
-        self.get_activity_attachment_by_id(result.last_insert_rowid())
-            .await
-    }
-
-    /// Get activity attachment by ID
-    pub async fn get_activity_attachment_by_id(&self, id: i64) -> Result<ActivityAttachment> {
-        let row = sqlx::query("SELECT * FROM activity_attachments WHERE id = ?")
-            .bind(id)
-            .fetch_one(&self.pool)
-            .await?;
-
-        self.row_to_activity_attachment(&row).await
-    }
-
-    /// Get attachments for an activity
-    pub async fn get_activity_attachments(
-        &self,
-        activity_id: i64,
-    ) -> Result<Vec<ActivityAttachment>> {
-        let rows = sqlx::query(
-            "SELECT * FROM activity_attachments WHERE activity_id = ? ORDER BY created_at DESC",
-        )
-        .bind(activity_id)
-        .fetch_all(&self.pool)
-        .await?;
-
-        let mut attachments = Vec::new();
-        for row in rows {
-            attachments.push(self.row_to_activity_attachment(&row).await?);
-        }
-
-        Ok(attachments)
-    }
-
-    /// Delete an activity attachment
-    pub async fn delete_activity_attachment(&self, id: i64) -> Result<()> {
-        sqlx::query("DELETE FROM activity_attachments WHERE id = ?")
-            .bind(id)
-            .execute(&self.pool)
-            .await?;
-        Ok(())
-    }
-
-    /// Helper method to convert database row to Activity struct
-    async fn row_to_activity(&self, row: &sqlx::sqlite::SqliteRow) -> Result<Activity> {
-        let category_str: String = row.try_get("category")?;
-        let category = category_str.parse::<ActivityCategory>()?;
-
-        let activity_data_str: Option<String> = row.try_get("activity_data")?;
-        let activity_data = activity_data_str
-            .map(|s| serde_json::from_str(&s))
-            .transpose()?;
-
-        let activity_date: DateTime<Utc> = row.try_get("activity_date")?;
-        let created_at: DateTime<Utc> = row.try_get("created_at")?;
-        let updated_at: DateTime<Utc> = row.try_get("updated_at")?;
-
-        Ok(Activity {
-            id: row.try_get("id")?,
-            pet_id: row.try_get("pet_id")?,
-            category,
-            subcategory: row.try_get("subcategory")?,
-            title: row.try_get("title")?,
-            description: row.try_get("description")?,
-            activity_date,
-            activity_data,
-            cost: row.try_get("cost")?,
-            currency: row.try_get("currency")?,
-            location: row.try_get("location")?,
-            mood_rating: row.try_get("mood_rating")?,
-            created_at,
-            updated_at,
-        })
-    }
-
-    /// Helper method to convert database row to ActivityAttachment struct
-    async fn row_to_activity_attachment(
-        &self,
-        row: &sqlx::sqlite::SqliteRow,
-    ) -> Result<ActivityAttachment> {
-        let file_type_str: String = row.try_get("file_type")?;
-        let file_type = file_type_str.parse::<ActivityAttachmentType>()?;
-
-        let metadata_str: Option<String> = row.try_get("metadata")?;
-        let metadata = metadata_str.map(|s| serde_json::from_str(&s)).transpose()?;
-
-        let created_at: DateTime<Utc> = row.try_get("created_at")?;
-
-        Ok(ActivityAttachment {
-            id: row.try_get("id")?,
-            activity_id: row.try_get("activity_id")?,
-            file_path: row.try_get("file_path")?,
-            file_type,
-            file_size: row.try_get("file_size")?,
-            thumbnail_path: row.try_get("thumbnail_path")?,
-            metadata,
-            created_at,
-        })
-    }
-
-    // Data Migration and Backup Utilities
-
-    /// Export all activities for a pet to JSON format for backup/portability
-    pub async fn export_activities_to_json(
-        &self,
-        pet_id: Option<i64>,
-    ) -> Result<serde_json::Value> {
-        let activities = if let Some(pid) = pet_id {
-            self.get_activities(Some(pid), None, None, None).await?
-        } else {
-            self.get_activities(None, None, None, None).await?
-        };
-
-        let mut export_data = serde_json::json!({
-            "export_version": "1.0",
-            "export_timestamp": Utc::now(),
-            "pet_id": pet_id,
-            "total_activities": activities.total_count,
-            "activities": []
-        });
-
-        for activity in activities.activities {
-            // Get attachments for this activity
-            let attachments = self.get_activity_attachments(activity.id).await?;
-
-            let activity_export = serde_json::json!({
-                "id": activity.id,
-                "pet_id": activity.pet_id,
-                "category": activity.category,
-                "subcategory": activity.subcategory,
-                "title": activity.title,
-                "description": activity.description,
-                "activity_date": activity.activity_date,
-                "activity_data": activity.activity_data,
-                "cost": activity.cost,
-                "currency": activity.currency,
-                "location": activity.location,
-                "mood_rating": activity.mood_rating,
-                "created_at": activity.created_at,
-                "updated_at": activity.updated_at,
-                "attachments": attachments
-            });
-
-            export_data["activities"]
-                .as_array_mut()
-                .unwrap()
-                .push(activity_export);
-        }
-
-        Ok(export_data)
-    }
-
-    /// Import activities from JSON format with validation and rollback support
-    pub async fn import_activities_from_json(
-        &self,
-        json_data: serde_json::Value,
-    ) -> Result<ImportResult> {
-        let mut import_result = ImportResult {
-            total_imported: 0,
-            total_failed: 0,
-            errors: Vec::new(),
-            rollback_data: Vec::new(),
-        };
-
-        // Validate import format
-        if !json_data.is_object() {
-            return Err(anyhow::anyhow!("Invalid import format: not an object"));
-        }
-
-        let activities = json_data["activities"]
-            .as_array()
-            .ok_or_else(|| anyhow::anyhow!("Invalid import format: missing activities array"))?;
-
-        // Begin transaction for atomic import
-        let mut transaction = self.pool.begin().await?;
-
-        for (index, activity_json) in activities.iter().enumerate() {
-            match self
-                .import_single_activity(&mut transaction, activity_json)
-                .await
-            {
-                Ok(activity_id) => {
-                    import_result.total_imported += 1;
-                    import_result.rollback_data.push(activity_id);
-                }
-                Err(e) => {
-                    import_result.total_failed += 1;
-                    import_result
-                        .errors
-                        .push(format!("Activity {}: {}", index + 1, e));
-                }
-            }
-        }
-
-        // Commit or rollback based on import success
-        if import_result.total_failed > 0 {
-            transaction.rollback().await?;
-            import_result.rollback_data.clear();
-            return Err(anyhow::anyhow!(
-                "Import failed with {} errors: {:?}",
-                import_result.total_failed,
-                import_result.errors
-            ));
-        }
-
-        transaction.commit().await?;
-        Ok(import_result)
-    }
-
-    /// Validate activity data consistency and integrity
-    pub async fn validate_activity_data(&self, pet_id: Option<i64>) -> Result<ValidationReport> {
-        let mut report = ValidationReport {
-            total_activities: 0,
-            valid_activities: 0,
-            issues: Vec::new(),
-            orphaned_attachments: 0,
-            missing_pets: Vec::new(),
-        };
-
-        // Get all activities or activities for specific pet
-        let activities = if let Some(pid) = pet_id {
-            self.get_activities(Some(pid), None, None, None).await?
-        } else {
-            self.get_activities(None, None, None, None).await?
-        };
-
-        report.total_activities = activities.total_count;
-
-        for activity in activities.activities {
-            let mut activity_valid = true;
-
-            // Validate pet exists
-            if self.get_pet_by_id(activity.pet_id).await.is_err() {
-                report.missing_pets.push(activity.pet_id);
-                report.issues.push(format!(
-                    "Activity {}: Referenced pet {} does not exist",
-                    activity.id, activity.pet_id
-                ));
-                activity_valid = false;
-            }
-
-            // Validate required fields
-            if activity.title.trim().is_empty() {
-                report
-                    .issues
-                    .push(format!("Activity {}: Empty title", activity.id));
-                activity_valid = false;
-            }
-
-            // Validate date is reasonable (not in far future or past)
-            let now = Utc::now();
-            let min_date = now - chrono::Duration::days(365 * 20); // 20 years ago
-            let max_date = now + chrono::Duration::days(365); // 1 year in future
-
-            if activity.activity_date < min_date || activity.activity_date > max_date {
-                report.issues.push(format!(
-                    "Activity {}: Suspicious date {}",
-                    activity.id, activity.activity_date
-                ));
-                activity_valid = false;
-            }
-
-            // Validate mood rating if present
-            if let Some(mood) = activity.mood_rating {
-                if !(1..=5).contains(&mood) {
-                    report.issues.push(format!(
-                        "Activity {}: Invalid mood rating {}",
-                        activity.id, mood
-                    ));
-                    activity_valid = false;
-                }
-            }
-
-            // Validate JSON data if present
-            if let Some(data) = &activity.activity_data {
-                if !data.is_object() {
-                    report.issues.push(format!(
-                        "Activity {}: Invalid activity data format",
-                        activity.id
-                    ));
-                    activity_valid = false;
-                }
-            }
-
-            if activity_valid {
-                report.valid_activities += 1;
-            }
-        }
-
-        // Check for orphaned attachments
-        let all_attachments = sqlx::query("SELECT COUNT(*) as count FROM activity_attachments WHERE activity_id NOT IN (SELECT id FROM activities)")
-            .fetch_one(&self.pool)
-            .await?;
-
-        report.orphaned_attachments = all_attachments.try_get::<i64, _>("count")?;
-
-        if report.orphaned_attachments > 0 {
-            report.issues.push(format!(
-                "Found {} orphaned attachments",
-                report.orphaned_attachments
-            ));
-        }
-
-        Ok(report)
-    }
-
-    /// Clean up orphaned data and fix consistency issues
-    pub async fn cleanup_activity_data(&self) -> Result<CleanupReport> {
-        let mut report = CleanupReport {
-            orphaned_attachments_removed: 0,
-            invalid_activities_fixed: 0,
-            fts_entries_rebuilt: 0,
-        };
-
-        // Remove orphaned attachments
-        let orphaned_result = sqlx::query(
-            "DELETE FROM activity_attachments WHERE activity_id NOT IN (SELECT id FROM activities)",
-        )
-        .execute(&self.pool)
-        .await?;
-
-        report.orphaned_attachments_removed = orphaned_result.rows_affected() as i64;
-
-        // Fix activities with empty titles
-        let fixed_titles = sqlx::query(
-            "UPDATE activities SET title = 'Untitled Activity' WHERE title = '' OR title IS NULL",
-        )
-        .execute(&self.pool)
-        .await?;
-
-        report.invalid_activities_fixed += fixed_titles.rows_affected() as i64;
-
-        // Fix invalid mood ratings
-        let fixed_moods = sqlx::query(
-            "UPDATE activities SET mood_rating = NULL WHERE mood_rating < 1 OR mood_rating > 5",
-        )
-        .execute(&self.pool)
-        .await?;
-
-        report.invalid_activities_fixed += fixed_moods.rows_affected() as i64;
-
-        // Rebuild FTS index
-        let _ = sqlx::query("DELETE FROM activities_fts")
-            .execute(&self.pool)
-            .await;
-
-        let activities =
-            sqlx::query("SELECT id, title, description, category, subcategory FROM activities")
-                .fetch_all(&self.pool)
-                .await?;
-
-        for row in activities {
-            let id: i64 = row.try_get("id")?;
-            let title: String = row.try_get("title")?;
-            let description: Option<String> = row.try_get("description")?;
-            let category: String = row.try_get("category")?;
-            let subcategory: String = row.try_get("subcategory")?;
-
-            sqlx::query("INSERT INTO activities_fts(rowid, title, description, category, subcategory) VALUES (?, ?, ?, ?, ?)")
-                .bind(id)
-                .bind(&title)
-                .bind(description.unwrap_or_default())
-                .bind(&category)
-                .bind(&subcategory)
-                .execute(&self.pool)
-                .await?;
-
-            report.fts_entries_rebuilt += 1;
-        }
-
-        Ok(report)
-    }
-
-    /// Create a backup of all activity data before major operations
-    pub async fn create_activity_backup(&self) -> Result<serde_json::Value> {
-        let backup_data = serde_json::json!({
-            "backup_version": "1.0",
-            "backup_timestamp": Utc::now(),
-            "activities": self.export_activities_to_json(None).await?,
-            "database_info": {
-                "total_activities": sqlx::query("SELECT COUNT(*) as count FROM activities")
-                    .fetch_one(&self.pool)
-                    .await?
-                    .try_get::<i64, _>("count")?,
-                "total_attachments": sqlx::query("SELECT COUNT(*) as count FROM activity_attachments")
-                    .fetch_one(&self.pool)
-                    .await?
-                    .try_get::<i64, _>("count")?
-            }
-        });
-
-        Ok(backup_data)
-    }
-
-    /// Restore activity data from backup with rollback support
-    pub async fn restore_from_backup(
-        &self,
-        backup_data: serde_json::Value,
-    ) -> Result<ImportResult> {
-        // Validate backup format
-        if backup_data.get("backup_version").is_none() {
-            return Err(anyhow::anyhow!("Invalid backup format: missing version"));
-        }
-
-        // Clear existing data in a transaction
-        let mut transaction = self.pool.begin().await?;
-
-        // Store current data for rollback
-        let current_backup = self.create_activity_backup().await?;
-
-        // Clear FTS first to avoid foreign key issues
-        sqlx::query("DELETE FROM activities_fts")
-            .execute(&mut *transaction)
-            .await?;
-
-        // Clear attachments first to avoid foreign key issues
-        sqlx::query("DELETE FROM activity_attachments")
-            .execute(&mut *transaction)
-            .await?;
-
-        // Clear activities
-        sqlx::query("DELETE FROM activities")
-            .execute(&mut *transaction)
-            .await?;
-
-        transaction.commit().await?;
-
-        // Import from backup
-        if let Some(activities_data) = backup_data.get("activities") {
-            match self
-                .import_activities_from_json(activities_data.clone())
-                .await
-            {
-                Ok(result) => Ok(result),
-                Err(_) => {
-                    // Rollback by restoring the current backup
-                    if let Some(current_activities) = current_backup.get("activities") {
-                        self.import_activities_from_json(current_activities.clone())
-                            .await
-                    } else {
-                        Err(anyhow::anyhow!(
-                            "Restore failed and rollback data is corrupted"
-                        ))
-                    }
-                }
-            }
-        } else {
-            Err(anyhow::anyhow!("Backup data does not contain activities"))
-        }
-    }
-
-    /// Helper method for importing a single activity (used in transactions)
-    async fn import_single_activity(
-        &self,
-        transaction: &mut sqlx::Transaction<'_, sqlx::Sqlite>,
-        activity_json: &serde_json::Value,
-    ) -> Result<i64> {
-        // Extract and validate fields
-        let pet_id = activity_json["pet_id"]
-            .as_i64()
-            .ok_or_else(|| anyhow::anyhow!("Missing or invalid pet_id"))?;
-
-        let category_str = activity_json["category"]
-            .as_str()
-            .ok_or_else(|| anyhow::anyhow!("Missing category"))?;
-        let category = category_str.parse::<ActivityCategory>()?;
-
-        let title = activity_json["title"]
-            .as_str()
-            .ok_or_else(|| anyhow::anyhow!("Missing title"))?
-            .to_string();
-
-        let subcategory = activity_json["subcategory"]
-            .as_str()
-            .unwrap_or("")
-            .to_string();
-
-        let description = activity_json["description"].as_str().map(|s| s.to_string());
-
-        let activity_date_str = activity_json["activity_date"]
-            .as_str()
-            .ok_or_else(|| anyhow::anyhow!("Missing activity_date"))?;
-        let activity_date =
-            chrono::DateTime::parse_from_rfc3339(activity_date_str)?.with_timezone(&Utc);
-
-        let activity_data = activity_json.get("activity_data").cloned();
-        let activity_data_json = activity_data
-            .map(|data| serde_json::to_string(&data))
-            .transpose()?;
-
-        let cost = activity_json["cost"].as_f64();
-        let currency = activity_json["currency"].as_str().map(|s| s.to_string());
-        let location = activity_json["location"].as_str().map(|s| s.to_string());
-        let mood_rating = activity_json["mood_rating"].as_i64().map(|r| r as i32);
-
-        let now = Utc::now();
-
-        // Insert activity
-        let result = sqlx::query(
-            r#"
-            INSERT INTO activities (pet_id, category, subcategory, title, description, activity_date, 
-                                  activity_data, cost, currency, location, mood_rating, created_at, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            "#
-        )
-        .bind(pet_id)
-        .bind(category.to_string())
-        .bind(&subcategory)
-        .bind(&title)
-        .bind(&description)
-        .bind(activity_date)
-        .bind(activity_data_json)
-        .bind(cost)
-        .bind(&currency)
-        .bind(&location)
-        .bind(mood_rating)
-        .bind(now)
-        .bind(now)
-        .execute(&mut **transaction)
-        .await?;
-
-        let activity_id = result.last_insert_rowid();
-
-        // Update FTS table
-        sqlx::query("INSERT INTO activities_fts(rowid, title, description, category, subcategory) VALUES (?, ?, ?, ?, ?)")
-            .bind(activity_id)
-            .bind(&title)
-            .bind(description.unwrap_or_default())
-            .bind(category.to_string())
-            .bind(&subcategory)
-            .execute(&mut **transaction)
-            .await?;
-
-        // Import attachments if present
-        if let Some(attachments) = activity_json["attachments"].as_array() {
-            for attachment_json in attachments {
-                self.import_single_attachment(&mut *transaction, activity_id, attachment_json)
-                    .await?;
-            }
-        }
-
-        Ok(activity_id)
-    }
-
-    /// Helper method for importing a single attachment
-    async fn import_single_attachment(
-        &self,
-        transaction: &mut sqlx::Transaction<'_, sqlx::Sqlite>,
-        activity_id: i64,
-        attachment_json: &serde_json::Value,
-    ) -> Result<i64> {
-        let file_path = attachment_json["file_path"]
-            .as_str()
-            .ok_or_else(|| anyhow::anyhow!("Missing file_path"))?
-            .to_string();
-
-        let file_type_str = attachment_json["file_type"]
-            .as_str()
-            .ok_or_else(|| anyhow::anyhow!("Missing file_type"))?;
-        let file_type = file_type_str.parse::<ActivityAttachmentType>()?;
-
-        let file_size = attachment_json["file_size"].as_i64();
-        let thumbnail_path = attachment_json["thumbnail_path"]
-            .as_str()
-            .map(|s| s.to_string());
-
-        let metadata = attachment_json.get("metadata").cloned();
-        let metadata_json = metadata
-            .map(|data| serde_json::to_string(&data))
-            .transpose()?;
-
-        let now = Utc::now();
-
-        let result = sqlx::query(
-            r#"
-            INSERT INTO activity_attachments (activity_id, file_path, file_type, file_size, thumbnail_path, metadata, created_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?)
-            "#
-        )
-        .bind(activity_id)
-        .bind(&file_path)
-        .bind(file_type.to_string())
-        .bind(file_size)
-        .bind(&thumbnail_path)
-        .bind(metadata_json)
-        .bind(now)
-        .execute(&mut **transaction)
-        .await?;
-
-        Ok(result.last_insert_rowid())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use chrono::{TimeZone, Utc};
-    use tempfile::NamedTempFile;
-
-    async fn setup_test_db() -> Result<super::super::PetDatabase> {
-        let temp_file = NamedTempFile::new()?;
-        let db =
-            super::super::PetDatabase::new_for_test(temp_file.path().to_str().unwrap()).await?;
-
-        // Create a test pet first
-        let pet_request = CreatePetRequest {
-            name: "Test Pet".to_string(),
-            species: PetSpecies::Dog,
-            breed: Some("Golden Retriever".to_string()),
-            gender: PetGender::Male,
-            birth_date: chrono::NaiveDate::from_ymd_opt(2020, 1, 1).unwrap(),
-            weight_kg: Some(25.5),
-            color: Some("Golden".to_string()),
-            notes: Some("Test pet for activity tests".to_string()),
-            photo_path: None,
-        };
-
-        db.create_pet(pet_request).await?;
-        Ok(db)
-    }
-
-    fn create_test_activity_request(pet_id: i64) -> ActivityCreateRequest {
-        ActivityCreateRequest {
-            pet_id,
-            category: ActivityCategory::Health,
-            subcategory: "Vaccination".to_string(),
-            title: "Annual Vaccination".to_string(),
-            description: Some("Annual vaccination at vet clinic".to_string()),
-            activity_date: Utc.with_ymd_and_hms(2024, 1, 15, 10, 30, 0).unwrap(),
-            activity_data: Some(serde_json::json!({
-                "veterinarian_name": "Dr. Smith",
-                "clinic_name": "Pet Care Clinic",
-                "symptoms": ["Routine checkup"],
-                "diagnosis": "Healthy",
-                "is_critical": false
-            })),
-            cost: None, // Avoid database type issues
-            currency: Some("USD".to_string()),
-            location: Some("Pet Care Clinic".to_string()),
-            mood_rating: Some(3),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_create_activity() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-
-        let activity = db.create_activity(request).await.unwrap();
-
-        assert_eq!(activity.pet_id, 1);
-        assert_eq!(activity.category, ActivityCategory::Health);
-        assert_eq!(activity.subcategory, "Vaccination");
-        assert_eq!(activity.title, "Annual Vaccination");
-        assert_eq!(activity.cost, None);
-        assert_eq!(activity.currency, Some("USD".to_string()));
-        assert_eq!(activity.mood_rating, Some(3));
-        assert!(activity.id > 0);
-    }
-
-    #[tokio::test]
-    async fn test_get_activity_by_id() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-
-        let created_activity = db.create_activity(request).await.unwrap();
-        let retrieved_activity = db.get_activity_by_id(created_activity.id).await.unwrap();
-
-        assert_eq!(created_activity.id, retrieved_activity.id);
-        assert_eq!(created_activity.title, retrieved_activity.title);
-        assert_eq!(created_activity.category, retrieved_activity.category);
-        assert_eq!(created_activity.pet_id, retrieved_activity.pet_id);
-    }
-
-    #[tokio::test]
-    async fn test_get_activities_with_pagination() {
-        let db = setup_test_db().await.unwrap();
-
-        // Create multiple activities
-        for i in 1..=5 {
-            let mut request = create_test_activity_request(1);
-            request.title = format!("Activity {i}");
-            db.create_activity(request).await.unwrap();
-        }
-
-        let result = db
-            .get_activities(Some(1), None, Some(3), Some(0))
-            .await
-            .unwrap();
-
-        assert_eq!(result.activities.len(), 3);
-        assert_eq!(result.total_count, 5);
-        assert!(result.has_more);
-    }
-
-    #[tokio::test]
-    async fn test_update_activity() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-
-        let created_activity = db.create_activity(request).await.unwrap();
-
-        let update_request = ActivityUpdateRequest {
-            title: Some("Updated Title".to_string()),
-            subcategory: Some("Updated Subcategory".to_string()),
-            description: Some("Updated description".to_string()),
-            ..Default::default()
-        };
-
-        let updated_activity = db
-            .update_activity(created_activity.id, update_request)
-            .await
-            .unwrap();
-
-        assert_eq!(updated_activity.title, "Updated Title");
-        assert_eq!(updated_activity.subcategory, "Updated Subcategory");
-        assert_eq!(
-            updated_activity.description,
-            Some("Updated description".to_string())
-        );
-        assert_eq!(updated_activity.id, created_activity.id);
-    }
-
-    #[tokio::test]
-    async fn test_delete_activity() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-
-        let created_activity = db.create_activity(request).await.unwrap();
-
-        // Verify activity exists
-        assert!(db.get_activity_by_id(created_activity.id).await.is_ok());
-
-        // Delete activity
-        db.delete_activity(created_activity.id).await.unwrap();
-
-        // Verify activity no longer exists
-        assert!(db.get_activity_by_id(created_activity.id).await.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_search_activities() {
-        let db = setup_test_db().await.unwrap();
-
-        // Create activities with searchable content
-        let mut request1 = create_test_activity_request(1);
-        request1.title = "Vet Checkup".to_string();
-        request1.description = Some("Annual health examination".to_string());
-        db.create_activity(request1).await.unwrap();
-
-        let mut request2 = create_test_activity_request(1);
-        request2.title = "Dog Walk".to_string();
-        request2.description = Some("Morning exercise in park".to_string());
-        db.create_activity(request2).await.unwrap();
-
-        // Search for "vet" should find the first activity
-        let result = db
-            .search_activities("vet", Some(1), Some(10), Some(0))
-            .await
-            .unwrap();
-
-        assert_eq!(result.activities.len(), 1);
-        assert_eq!(result.activities[0].title, "Vet Checkup");
-    }
-
-    #[tokio::test]
-    async fn test_different_activity_categories() {
-        let db = setup_test_db().await.unwrap();
-
-        // Create activities of different categories
-        let categories = [
-            ActivityCategory::Health,
-            ActivityCategory::Growth,
-            ActivityCategory::Diet,
-            ActivityCategory::Lifestyle,
-            ActivityCategory::Expense,
-        ];
-
-        for category in categories.iter() {
-            let mut request = create_test_activity_request(1);
-            request.category = *category;
-            request.title = format!("{category:?} Activity");
-            db.create_activity(request).await.unwrap();
-        }
-
-        let result = db.get_activities(Some(1), None, None, None).await.unwrap();
-        assert_eq!(result.activities.len(), 5);
-
-        // Verify all categories are represented
-        let found_categories: std::collections::HashSet<ActivityCategory> =
-            result.activities.iter().map(|a| a.category).collect();
-        assert_eq!(found_categories.len(), 5);
-    }
-
-    #[tokio::test]
-    async fn test_activity_with_json_data() {
-        let db = setup_test_db().await.unwrap();
-
-        let mut request = create_test_activity_request(1);
-        request.activity_data = Some(serde_json::json!({
-            "weight": {"value": 25.5, "unit": "kg"},
-            "height": {"value": 65, "unit": "cm"},
-            "milestone_type": "Adult Weight Reached",
-            "development_stage": "Adult"
-        }));
-
-        let activity = db.create_activity(request).await.unwrap();
-
-        assert!(activity.activity_data.is_some());
-        let data = activity.activity_data.unwrap();
-        assert_eq!(data["weight"]["value"], 25.5);
-        assert_eq!(data["weight"]["unit"], "kg");
-    }
-
-    #[tokio::test]
-    async fn test_create_activity_attachment() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-        let activity = db.create_activity(request).await.unwrap();
-
-        let attachment = db
-            .create_activity_attachment(
-                activity.id,
-                "/path/to/photo.jpg".to_string(),
-                ActivityAttachmentType::Photo,
-                Some(1024000),
-                Some("/path/to/thumbnail.jpg".to_string()),
-                Some(serde_json::json!({"camera": "iPhone 12", "location": "Home"})),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(attachment.activity_id, activity.id);
-        assert_eq!(attachment.file_path, "/path/to/photo.jpg");
-        assert_eq!(attachment.file_type, ActivityAttachmentType::Photo);
-        assert_eq!(attachment.file_size, Some(1024000));
-        assert!(attachment.id > 0);
-    }
-
-    #[tokio::test]
-    async fn test_get_activity_attachments() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-        let activity = db.create_activity(request).await.unwrap();
-
-        // Create multiple attachments
-        for i in 1..=3 {
-            db.create_activity_attachment(
-                activity.id,
-                format!("/path/to/photo{i}.jpg"),
-                ActivityAttachmentType::Photo,
-                Some(1024000 * i),
-                None,
-                None,
-            )
-            .await
-            .unwrap();
-        }
-
-        let attachments = db.get_activity_attachments(activity.id).await.unwrap();
-
-        assert_eq!(attachments.len(), 3);
-        for attachment in &attachments {
-            assert_eq!(attachment.activity_id, activity.id);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_delete_activity_attachment() {
-        let db = setup_test_db().await.unwrap();
-        let request = create_test_activity_request(1);
-        let activity = db.create_activity(request).await.unwrap();
-
-        let attachment = db
-            .create_activity_attachment(
-                activity.id,
-                "/path/to/photo.jpg".to_string(),
-                ActivityAttachmentType::Photo,
-                Some(1024000),
-                None,
-                None,
-            )
-            .await
-            .unwrap();
-
-        // Verify attachment exists
-        assert!(db
-            .get_activity_attachment_by_id(attachment.id)
-            .await
-            .is_ok());
-
-        // Delete attachment
-        db.delete_activity_attachment(attachment.id).await.unwrap();
-
-        // Verify attachment no longer exists
-        assert!(db
-            .get_activity_attachment_by_id(attachment.id)
-            .await
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn test_activity_data_integrity() {
-        let db = setup_test_db().await.unwrap();
-
-        // Test activity with all optional fields None
-        let mut minimal_request = create_test_activity_request(1);
-        minimal_request.description = None;
-        minimal_request.activity_data = None;
-        minimal_request.cost = None;
-        minimal_request.currency = None;
-        minimal_request.location = None;
-        minimal_request.mood_rating = None;
-
-        let minimal_activity = db.create_activity(minimal_request).await.unwrap();
-
-        assert_eq!(minimal_activity.description, None);
-        assert_eq!(minimal_activity.activity_data, None);
-        assert_eq!(minimal_activity.cost, None);
-        assert_eq!(minimal_activity.currency, None);
-        assert_eq!(minimal_activity.location, None);
-        assert_eq!(minimal_activity.mood_rating, None);
-
-        // Verify required fields are still present
-        assert!(!minimal_activity.title.is_empty());
-        assert_eq!(minimal_activity.pet_id, 1);
-        assert_eq!(minimal_activity.category, ActivityCategory::Health);
-    }
-
-    #[tokio::test]
-    async fn test_activity_filtering_by_category() {
-        let db = setup_test_db().await.unwrap();
-
-        // Create activities of different categories
-        let mut health_request = create_test_activity_request(1);
-        health_request.category = ActivityCategory::Health;
-        health_request.title = "Health Activity".to_string();
-        db.create_activity(health_request).await.unwrap();
-
-        let mut diet_request = create_test_activity_request(1);
-        diet_request.category = ActivityCategory::Diet;
-        diet_request.title = "Diet Activity".to_string();
-        db.create_activity(diet_request).await.unwrap();
-
-        // Test filtering by specific category
-        let filters = ActivityFilters {
-            categories: Some(vec![ActivityCategory::Health]),
-            ..Default::default()
-        };
-
-        let result = db
-            .get_activities(Some(1), Some(filters), None, None)
-            .await
-            .unwrap();
-
-        // Note: The current implementation doesn't fully support filtering yet
-        // This test verifies the structure is in place
-        assert!(!result.activities.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_concurrent_activity_operations() {
-        let db = setup_test_db().await.unwrap();
-
-        // Test concurrent-style operations but in a controlled manner
-        // First create one activity to ensure the database is properly initialized
-        let init_request = create_test_activity_request(1);
-        db.create_activity(init_request).await.unwrap();
-
-        // Now create the rest sequentially to simulate concurrent success without race conditions
-        for i in 2..=5 {
-            let mut request = create_test_activity_request(1);
-            request.title = format!("Concurrent Activity {i}");
-            db.create_activity(request).await.unwrap();
-        }
-
-        // Verify we have 5 activities
-        let all_activities = db.get_activities(Some(1), None, None, None).await.unwrap();
-        assert_eq!(all_activities.activities.len(), 5);
-
-        // Verify they have different titles (simulating successful concurrent operations)
-        let titles: Vec<String> = all_activities
-            .activities
-            .iter()
-            .map(|a| a.title.clone())
-            .collect();
-        assert!(titles.iter().any(|t| t.contains("Activity")));
-    }
-}
