+++ src-tauri/src/errors/mod.rs
@@ -1,0 +1,5 @@
+pub mod common;
+pub mod pet;
+
+pub use common::*;
+pub use pet::*;
