+++ src-tauri/src/lib.rs
@@ -5,6 +5,7 @@ pub mod errors;
 pub mod logger;
 pub mod photo;
 pub mod protocol;
+pub mod validation;
 
 use commands::*;
 use tauri::http::Response;
@@ -32,19 +33,6 @@ pub fn run() {
             update_pet,
             delete_pet,
             reorder_pets,
-            // Activity management commands
-            create_activity,
-            get_activities,
-            get_activity_by_id,
-            update_activity,
-            delete_activity,
-            search_activities,
-            // Activity attachment commands
-            upload_activity_attachment,
-            upload_activity_attachment_from_path,
-            get_activity_attachments,
-            get_activity_attachment_by_id,
-            delete_activity_attachment,
             // Photo management commands
             upload_pet_photo,
             upload_pet_photo_from_path,
