+++ src-tauri/src/validation/mod.rs
@@ -1,0 +1,3 @@
+pub mod pet;
+
+pub use pet::*;
